+++ workspace/apps/server/src/controllers/account.ts
@@ -248,13 +248,16 @@ export const updateInfoMutation = procedure //
         try {
             const { session, role, DB } = options.ctx;
             if (session.data.profile) {
-                const profile = await options.ctx.DB.profile.update({
+                const profile = await DB.profile.update({
                     where: {
                         id: session.data.profile.id,
                         deleted: false,
                     },
                     data: {
-                        avatar: options.input.avatar !== undefined ? options.input.avatar : null,
+                        avatar:
+                            options.input.avatar !== undefined //
+                                ? options.input.avatar
+                                : undefined,
                     },
                 });
                 return {
@@ -314,6 +317,10 @@ export const changePasswordMutation = procedure //
                 id: number;
                 role: AccessorRole;
                 password: string;
+                token: {
+                    id: number;
+                    version: number;
+                };
             } = await (async () => {
                 switch (payload.data.role) {
                     /* 用户 */
@@ -328,6 +335,12 @@ export const changePasswordMutation = procedure //
                                 id: true,
                                 role: true,
                                 password: true,
+                                token: {
+                                    select: {
+                                        id: true,
+                                        version: true,
+                                    },
+                                },
                             },
                         });
                         return user;
@@ -344,6 +357,12 @@ export const changePasswordMutation = procedure //
                                 id: true,
                                 role: true,
                                 password: true,
+                                token: {
+                                    select: {
+                                        id: true,
+                                        version: true,
+                                    },
+                                },
                             },
                         });
                         return staff;
@@ -388,15 +407,14 @@ export const changePasswordMutation = procedure //
                 }
 
                 /* 吊销令牌 */
-                const token = options.ctx.session.data.token;
-                token.version++;
-                tokens.set(token.id, token.version);
+                account.token.version++;
+                tokens.set(account.token.id, account.token.version);
                 await options.ctx.DB.token.update({
                     where: {
-                        id: token.id,
+                        id: account.token.id,
                     },
                     data: {
-                        version: token.version,
+                        version: account.token.version,
                     },
                 });